Attempt to shut up the compiler.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@11928 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/PATM/PATMSSM.cpp
@@ -591,7 +591,7 @@ DECLCALLBACK(int) patmr3Load(PVM pVM, PSSMHANDLE pSSM, uint32_t u32Version)
             AssertRCReturn(rc, rc);
 
             /* rec.pRelocPos now contains the relative position inside the hypervisor area. */
-            offset = (int32_t)rec.pRelocPos;
+            offset = (int32_t)(int64_t)rec.pRelocPos;
             /* Convert to HC pointer again. */
             PATM_ADD_PTR(rec.pRelocPos, pVM->patm.s.pPatchMemHC);
             pFixup = (RTRCPTR *)rec.pRelocPos;
